backend/requirements.txt
tensorflow==2.10.1
keras==2.10.0
opencv-python==4.8.0.76
numpy==1.23.5
matplotlib==3.7.1
Pillow==9.5.0
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-multipart==0.0.6
scikit-learn==1.3.0
python-dotenv==1.0.0